Updated client logging to apply details via MDC Added support to update how glasspane dimensions are determined Added support to override the FontAwesome URL Added ability to specify GoogleFonts are already loaded

+++ gui/src/main/scala/io/youi/BackgroundUpdates.scala
@@ -3,5 +3,5 @@ package io.youi
 import org.scalajs.dom.window
 
 object BackgroundUpdates extends UpdateSupport {
-  override protected def run(): Unit = window.setTimeout(() => updateFunction(System.currentTimeMillis()), 100)
+  override protected def run(): Unit = window.setTimeout(() => updateFunction(System.currentTimeMillis().toDouble), 100)
 }
